Changes to peer offline handling
The current offline handling is a too agressive and needs to be
considered a bit more carefully.

This change removes the check in ConnectionManager to immediately fail
dials if the peer is marked as offline. This did not belong in the
connection manager anyway, and more intelligent handling of this case
will live higher up in the stack.

Nodes will still not broadcast to offline peers however direct send
messages will still attempt to dial. This should solve the issue on
mobile where the phone does not attempt to connect to a base node it
cannot talk to.

Nodes will no longer mark all their base node neighbours as offline. If
a certain percentage (30%) of neighbouring nodes are already marked as
offline, no further nodes will be marked as offline.

Added `reset-offline-peers` command

A more sophisticated solutions may involve:
- Somehow tell comms that transport-level connectivity has been lost
  (probably requires the OS to tell comms about the state of internet connectivity)
- A node should ask neighbours and come to a consensus on the state of
  their neighbouring peer list.