Update Tx Service to one-shot response for SendDirect(..) & fix OMS bug
This PR updates the Tx Service to use the embedded oneshot reply channel to determine if a Direct message was sent or not rather than relying on the MessagingEvent stream.

It also fixes a bug in the Output Manager service where an Inbound transaction output was not fully encumbered so if the app restarted it would be cleared as a short term encuberance.

Finally some Clippy cleanup.