Add UTXO selection strategy for large txs
The default UTXO selection strategy is to use the smallest outputs
first. This works well most of the time, but for really large trans-
action amounts, we can run into the transaction input limits.

This PR adds a new strategy and a selection heuristic for selcring it.
`Largest` does what it says on the box. It's the inverse of `Smallest`.

If the total transaction amount > alpha * My largest UTXO, then we
select Largest (for now alpha = 1).

The coin_split strategy also uses Largest (since that makes the most
sense if you're trying to make more UTXOs)