Reset failed connection attempts when discovery/join received
A peer's offline state would be removed when receiving a discovery/join
however, it's failed connection attempt count was not making it possible
for a single failed connection attempt to mark the peer as offline.

This PR ensures that a spurious connection would not cause a peer to
mistakenly be marked as offline.