Add FFI function to return list of cancelled transactions
This PR adds a new FFI function to return all cancelled transactions. This list includes cancelled Pending Inbound and Outbound transactions that are converted to CompletedTransaction structs for the final collection so that all three transaction types can live in the same list.

To facilitate this the database is updated to use a separate flag to indicate if a transaction (inbound, outbound and completed) is cancelled while maintaining the original status of the transactions.