Improves broadcast and propagation strategies
Propagation didn't include connected peers in propagation.
This could cause messages to stay within their neighbourhood cluster in
the XOR metric space.
This PR includes all connected peers as propagation candidates, then
selects a subset of peers that are closer to the message destination.
If no closer peers can be found (and the node is well connected),
the propagation ends.

A distinction is made between propagating a message along and
broadcasting a new message. Broadcasted messages are sent to neighbours
for propagation. Propagated messages are forwarded closer to the
message destination. If a destination is not provided, the message is
broadcast to a subset of connected nodes and neighbours.

- Added `BroadcastStrategy::Propagate` that adds connected nodes to the
  propagation set for consideration
- Propagation selects a subset peers that are closer to the destination
- Renamed `neighbours` to `broadcast` in the outbound requester
  interface
- Use `NodeId` instead of `CommsPublicKey` for excluded peer list (for conistency
  and it has a much smaller memory footprint)
- Updated memorynet example to test domain-level message propagation