Close the connection earlier if a peer is banned
If a peer was banned, the identity protocol would complete before the connection
was closed. This means that the peer connection of a dialing peer
would complete successfully before being immediately closed by the
listener. The banned dialing peer would retry indefinitely without
giving up.

This PR adds a check to the listener if the peer is banned _before_
beginning the peer identity protocol. This reduces the amount of wasted
work the listener has to do as well failing the dialer's peer connection.
The dialer will treat this as any other kind of failure and mark the
peer as offline.