Added multiplexed substream counts
Wrapped yamux structs to keep track of the number of substreams
that are currently being used. This will be used when doing connection
reaping. The connectivity manager will use this value to determine if it
can safely close connections.

TODO: The messaging protocol must close the peer's substream if there is
a long enough period of inactivity. The inbound messaging handler was
refactored into a separate module to keep it consistent with the
outbound handler.