Add `get_cancelled_transaction_by_id` to FFI & update cancel callback
This PR add a `get_cancelled_transaction_by_id(…)` function to the FFI to request a single cancelled transaction. It also updates the `callback_transaction_cancellation(…)` callback in the FFI to return a pointer to the Transaction that was cancelled.

To facilitate this the Transaction service backend was updated to allow for queries of cancelled inbound, outbound and completed transactions.