Add local pub key to the source and dest fields for cancelled txs in FFI
The cancelled transactions returned by the `wallet_get_cancelled_transactions(…)` FFI function returns a list of CompletedTransactions of which some could be previous PendingInbound and PendingOutbound transactions. These previously pending transactions done have destination and source public keys in them because they are destined for or originate from the local node. In the conversion to a CompletedTransaction the respective pub key was zeroed.

This PR updated the  `wallet_get_cancelled_transactions(…)` function to include the local pub key in the respective fields.