Messaging rate limiting
This PR implements a permit-based rate limiter on the inbound
messaging stream that limits the amount of inbound messages
that will be read off the stream in a given window.

Given, `capacity: usize` and `restock_interval: Duration`, up to `capacity` messages can be received
within the `restock_interval`.  If the capacity is exhaused, backpressure will apply to the sender.
The `capacity` is set to 10, and the `restock_interval` is set to 100ms and these apply _per peer_
(+100 messages per second).