Increased redundancy for SAF messages
When a message (with an explicit destination) is stored by a node,
broadcast `n` messages to connected peers that are close to the
destination to increase redundancy.

As a message is propagated closer to the destination, the message
will start to be stored for an offline peer once it reaches the
neighbourhood of the destination. Because propagation is more targeted,
it may miss some nodes that could be considered neighbours by the
wallet. This PR increases the number of nodes that see the message
as the message reaches the neighbourhood.