Fix missing Source and Destination Pubkeys in FFI callback Tx return
For the CompletedTransaction returned by the transaction cancellation callback it is possible that the transaction is a cancelled Pending Incoming or Pending Outgoing transaction which does not include the destination and source pub keys respectively. This PR adds in the code to add in the missing pub key to the returned CompletedTransaction.