Refactor the Receive Transaction Protocol into the Protocol pattern
This PR refactors the Receive Transaction Protocol in the Transaction Service. It moves all the logic for handling accepting a transaction, sending the reply and waiting for the finalised transaction into a self contained Protocol pattern. This protocol was added to the Transaction Service and plumbed in as with the previous protocols.

All the send operations in the Send and Receive transaction protocols were updated to follow the strategy of attempt direct and only send via SAF if discovery is needed or the direct send times out.

It was also decided that the Send and Receive transaction protocols should not monitor the liveness of the counterparty for resending purposes so that was removed from the Send Transaction Protocol. This PR also removes the `test_resend_of_tx_on_pong_event` which tested this functionality.

The Send Transaction Protocol was also cleaned up to better reflect the stages it can be started in.

A test was added to test that both the send and receive transaction protocols restart when the wallet starts up again.