Resolve peer DNS addresses with Tor if it is available
This PR adds DNS address resolution using the Tor proxy if it is
available. This is a privacy enhancing change that applies to nodes
running in TCP mode with tor support. Previously, connecting to a peer's
DNS address would use the system's DNS resolution and that can be
exploited to deanonymise users.

- Use `DnsResolver` trait in `TcpTransport` to resolve addresses
- `TorDnsResolver` implementation that connects to the Tor proxy to
  resolve DNS requests
- `SystemDnsResolver` asynchronously (blocking thread) resolves DNS
  requests if Tor is not configured.
- Fix for the `tor_resolve` command in the SOCKS5 client
- Ensure that the node identity is correctly set to the public address
  specified in the node config, if applicable.