Remove Liveness and add a Low Power mode to the Wallet
In order to help reduce battery consumption this PR removes the Liveness service from the Wallet which is no longer required due to updates to the connectivity manager. 

This PR also adds two FFI functions to set Low power and Normal power modes. Currently these modes will change the Timeouts used by the two  base node monitoring protocols to poll the base node from 30 seconds to 5 minutes. The mobile client can switch to the low power mode when the app moves to the background when it doesn’t need to be as responsive and it will generate 10 times less network traffic and thus use less battery.