Made messaging pipeline optional in CommsBuilder
Messaging pipeline is optional. If a MessageingPipeline is not specified
in the builder, the messaging protocol, associated channels and tokio tasks
will never be initialized/spawned. This allows a more light-weight comms to be
initialized in tests that do not require the messaging pipeline (such as
integration tests for components using e.g. RPC only) without the added
cruft of setting up a _no-op_ messaging pipeline.