Have the Wallet backend persist the Comms Secret Key
Previously the frontend client of the Wallet FFI library was responsible for persisting the Comms Secret Key. This made backing up and restoring the wallet tricky and as such it was decided that this private key should be persisted in the backend sqlite db.

This PR updates wallet backend DB with a new table called `wallet_settings` which will be a key-value store of wallet settings. The FFI function `comms_config_create(…)` to first check the wallet backend to see if a Comms Private Key is stored. If it is then it is used else a new one is generated.  The `wallet_create(…)` function will write whatever Comms Private Key is provided to it to the db for persistence.

In order to facilitate the migration to this new setup a new function is added to the FFI, `comms_config_set_secret_key(…)` which allows the frontend to specify the Comms Secret Key used. When migrating the Comms Config will be created. This will generate a new key because none will be found in the backend and then the frontend persisted Secret key will be set using the `comms_config_set_secret_key(…)` function and there after the backend will persist this secret key.