Check in Cargo.lock
It's Rust best practice to check Cargo.lock into VCS for applications.
Now that the tari base node is the main project in this codebase, we
should follow that best practice.

This should reduce / eliminate hard-to-diagnose bugs due to different
developers running slightly different codebases due to differences in
dependent packages.