Github action for building base node binaries
Added a new github action for building tari_base_node binaries.
The action supports Ubuntu for now, with multiple targets (in progress).
The resulting binaries are zipped, hashed and copied to an S3 bucket.

The AVX2 feature flag is consistently applied acrosss the project to
support a build that does not support it.