Add partial wallet backup for unencrypted wallet backup
This PR adds functionality to produce a partial wallet backup that excludes the Comms Secret Key from the backup. This is to be used when the balance in a wallet is below a certain threshold so is low risk to backup without passphrase encryption. 

This use case is to allow for automatic backing up in the early use of the wallet while there are little funds so we don’t want to bother the nascent user with passphrase when the risk is low. When the funds are above a certain threshold an encryption passphrase would be required.

This PR adds:
- DB migration to store transaction direction for CompletedTransactions rather than relying on examining the Source Public Key. This is so that after a backup where the Private Key is cleared event though a new private key is generated the wallet can still tell which transactions were inbound or outbound.
- Functionality to perform the backup, clear the sensitive data and write it to the specified file.
- FFI interface to this functionality.

Restoration will be done by copying this backup file to the device and starting the wallet providing this db file as the database.

Because it is a Sqlite file and has the migration history inside it will automatically be migrated to new DB schema. When it is detected that there is no Comms Secret Key in the database a new one will be generated.