Fix bug where the failed UTXO Sync callback is called twice
Currently when the output manager service performs a UTXO validation if the process times out two events are emitted: one for the timeout and 1 for the failed protocol as by default only 1 attempt was made. Both these events trigger a callback to the FFI client. 

This PR stops the emission of one of these events on a Validation protocol timeout so only 1 callback occurs.

This PR also updates some of the logging in the Output Manager Service.