Remove unnecessary Options from StatusInfo
This is a quick cleanup PR to remove unnecessary Options from the
StateInfo. A height of None is equivalent to a height of 0 and
sync_peers = None is equivalent to an empty Vec. This PR removes
these extraneous Options from StateInfo.

Additionally, a possible bug vector is removed: if no blocks are requested
(block_nums.is_empty()), the request is not made at all. This bug
existed in horizon sync, where a peer would be banned for returning an empty
response even though the node requested zero items.