Remove Peer persistence in Wallet
The persistence of base node peers is a vestigial feature from a time where we though the Wallet backend would handle all of the base node address monitoring. It has turned out that the Clients will manage base node peers and provide them to the wallet during Runtime. Further more this db implementation was producing a bug in the mobile clients due to its lack of a db transaction between reading the db and deleting all the peers on startup. 

This PR removes the unneeded Peer persistence to get rid of this bug. 

Some extra logging is added to the wallet and wallet_ffi for better future debugging purposes.