Another fix for flaky rate_limit_restock
`rate_limit_restock` is flaky on very slow systems because
when more than one future is ready at the same time
`futures::select!` simply picks a _random_ future. This can mean that
restocking doesnt happen in some rare cases.

This PR changes the tests to use `future::select(A, B)` which polls A
then B. If A is the rate limiter and B is the test timer, then A will
always be polled first allowing it to restock.

Due to high resource contention and the time-sensitive nature of the
test, it is still conceivable for more than 50ms to expire between restocks
so the test will now strictly only test restock (i.e that at least one restock happens)