Add partial wallet backup using only file paths
The previous feature that allowed for a partial wallet backup (excluding the Comms Private Key) used an instantiated wallet instance to get the path of its database. This turned out to be unwieldy and it was requested that this feature be updated to just accept the file paths to the database file being backed up and to the backup location.

This PR replaces the `wallet_partial_backup(…)` FFI function with the `file_partial_backup(…)` which will copy the wallet database file and clear the Comms Private Key for unencrypted backup.