Update FFI wallet_get_pending_outbound_transactions(…) to work after a partial wallet restore
Previously this function compared the current Comms Public Key to the Source public key of each transaction to decide if it was outbound. After a Partial wallet restore a new Comms Public Key is generated so this function would not be able to tell that a given transaction was Outbound. 

This PR updates this function to use the stored TransactionDirection to make this determination which will work after a partial wallet restore.