Add remote coinbase transaction handling to the wallet
This PR adds the functionality for an external application to request a complete Coinbase transaction from the wallet by providing the reward, fees and block height at which this coinbase will appear. The Transaction Service will then request a coinbase key from the Output Manger Service which will generate a private key from the “coinbase” branch of the current Key Manager using the Block Height as an index. The transaction service will then produce and return the Coinbase transaction to the requester and start monitoring the chain for the Coinbase output to be mined. If the chain advances to a height beyond the Coinbase transactions block height and it has not appeared in the chain it is cancelled.
If multiple requests are made for a coinbase transaction of the same height the previous transaction and outputs are cancelled and the new output and transaction becomes the active Coinbase transaction for that block height.