Rollback on pruned sync validation failure
This PR addsan `EffectivePrunedHeight` chain metadata entry and
`InProgressPrunedSyncState` metadata item to the blockchain db.
`InProgressPrunedSyncState` is constructed and persisted once pruned sync
starts and is used to keep track of the start index of "temporary"
checkpoints created during sync.

If a horizon state validation error occurs the sync data is rolled back
by fetching each temporary checkpoint and reversing the operations and
deleting the checkpoints.

If horizon sync is interrupted, the state is loaded again and the sync is resumed.

One the sync is complete, the InProgressPrunedSyncState is deleted and
the chain metadata updated accordingly.

Other changes:

- Sync peers now contains each peer's chain metadata so that an appropriate
  peer can be selected for block sync i.e. a peer with an effective
  pruned height lower than the syncing nodes current height.
- Header sync does not reset the chain metadata to genesis anymore as a
  that should be in pruned sync.
- Prefixed all database functions that ony are applicable for horizon
  sync with `horizon_sync_` to indicate that they are specialized and
  shouldn't used in other areas of the system at this time.
- DRY'd up all the checkpoint "utility" functions. It is no longer
  required to maintain separate functions for lmdb and memorydb.
- Removed `n` clones when storing a block
- Added `set_metadata` and `delete_metadata` convinience calls to
  `DbTransaction`
- Optimised chain balance validation to not have to hold every
  kernel/UTXO in memory while summing commitments.
- A few error improvements and migration to `thiserror`
- New basic test for horizon sync rollback