Implement the interface for `tari_console_wallet` application
This PR implements the layout and UX of the `tari_console_wallet`. To run the wallet in the console type `cargo run --bin tari_console_wallet` in the root of the repo.

The data in this PR is currently all dummy data to demonstrate the layout and UX of the app. The next PR will be to plumb a full LibWallet backend into the app.

This PR also adds support to the OutputManagerService for an optional Time Locked balance if a chain tip is provided

Missing UI features to be added in later PRs are:
- Fully console to handle commands like those in the `tari_base_node` app