Add extra validation to `wallet_get_pending_X_transaction_by_id` methods
This PR adds an extra validation check to the `wallet_get_pending_outbound_transaction_by_id` and `wallet_get_pending_inbound_transaction_by_id` FFI methods. Previously if you called these functions with the TxId of an inbound/outbound completed transaction, respectively, that was not mined it would be mistakenly returns as the direction of the function that was called. The extra validation ensures that the transaction will not be returned unless the direction is strictly correct.

This should not have been an issue as the functions that return the TxIds for pending transactions that would be used to call these functions are correct but this validation makes sure this error is not possible.