Fix flakey ` test_output_manager_XXX` tests 🤞
The various `test_output_manager_XXX` tests have been flakey and I think this PR fixes the issue. Though its hard to test so we will need to see.

The issue seemed to be that there was a small chance that the first Unspent Output ended up being the same as the first output from the first Pending Transaction so when it was invalidated it still left an output with the same spending key in the db to be found when it shouldn’t.

The removed invalidation was not actually testing anything the rest of the test doesn’t test so was superfluous.