Simplified messaging protocol, fix for excessive delays
This PR simplifies the outbound messaging and allows it to use
`StreamExt::forward` instead of a custom `Stream` impl.

This simplification means that the MessageSent event is no longer
sent and that in some rare cases a message may be "marked" as
sent even if the peer will not receive the message (e.g. sudden disconnect)

It should be acknowleged that previously this was the case anyway,
and "successfully sent" is more accurately "successfully buffered".

- Removed MessageSent event
- Removed retries from protocol
- Outbound messaging will retry once to connect and establish a
  substream, this is to handle the case where a pending dial is cancelled
  due to tie breaking - so one and only one retry is required.
- Retries happen per peer and not per message.
- A fix for the delay that is experienced in some cases.
- Ensure that sinks are closed if the workers exit for any reason.
- Added messaging stress test