Add support for optional rolling log files to FFI wallet library
Currently LibWallet accepts a single file path for logging and write all logs to that file. This PR adds the ability to optionally specify whether you want the logs to use multiple rolling files and at what size of file to trigger the roll over. If non zero values are provided for both `num_rolling_log_files` and `size_per_log_file_bytes` then the fixed window rolling functionality of log4rs will be used. If this is not desired 0 values should be passed for these arguments.

The rolled over archived files will be numbered version of the original filename. For example if the logging file path that is provided is `/User/me/wallet/log_170820_14382455.log`, 2 roll over files are specified and the size threshold is 10_000 bytes. 
1. The most current log file will always be `log_170820_14382455.log` 
2. When that reaches 10_000 bytes in size it will be renamed to `log_170820_14382455.0.log` and logging will continue in `log_170820_14382455.log`.
3. When the next roll over limit is reached: 
    1. `log_170820_14382455.0.log` will be renamed to `log_170820_14382455.1.log` 
    2. the current `log_170820_14382455.log` will be renamed to `log_170820_14382455.0.log`
    3. logging will continue in `log_170820_14382455.log`.
4. When the next roll over occurs because we specified only 2 roll over files the oldest one (index 1) will be deleted, 0 will be renamed to 1 and the current file renamed to index 0.

This can be used to limit the amount of logs LibWallet generates per session.