** Adjust input and output definitions
** Update tari_crypto to v0.7.0
** Update tari_utilities version

** Add script_hash to inputs and outputs

Add the additional script hash field to inputs and outputs.
Specifically, this entails

* Adding the hash itself (Vec<u8>) to the blinded versions of the
structs (TransactionInput and TransactionOutput).
* Adding the `TariScript` object itself to `UnblindedOutput`, since this
represents the "opened" version of the output.
* Adding various `From` and `TryFrom` impls
* Adapting the protobug definitions where necessary.
* Changing all constructors to use the DEFAULT script (for now).

I also made the Input and Output structs immutable objects, as they
should be, by making all the fields private, adding the missing
accessors and fixing all calls that accessed the fields directly.
Luckily, there were no mutable instances of the structs, as this would
likely have indicated a bug somewhere, so it was a simple refactor.