Add script hash features to UnblindedOutput
In theory, we can leave `TransactionOutput` and `Transactioninput` as is
and the consensus rules will still work even after applying the Tari
script changes, as long as..

`UnblindedTransactionOutput` produces the adjusted commitment in a
fairly seamless manner. This set of changes tries to do just that by
requiring the use of the `commitment` method to calculate a UTXO
commitment from an unblinded output.

At this point tests will be failing because the kernel signature has not
been updated accordingly, but it's a step in the right direction.